backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
